add immutable
This commit is massive - it replaces every (sensible) instance of a javascript object or javascript array with an immutable map or an immutable list, respectively. Some untouched objects are dispatch actions, as these need to be plain objects. Also untoucned was and request body. no need to do it there