swaps search in header for new component
This change implements the new search component into the system. now a user can submit a query via the search bar in the header, which sends a get request to the searches api index endpoint. this returns a list of books which are saved in state, and the user is redirected to "/search"